requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pypdf>=4.0.0
openpyxl>=3.1.0
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
xmltodict>=0.13.0
requests>=2.31.0
python-dotenv>=1.0.0
